Refactor to move new healing items to their own directory
