+++ src/main/java/io/github/pylonmc/pylon/base/PylonItems.java
@@ -5,6 +5,7 @@ import io.github.pylonmc.pylon.base.items.Fiber;
 import io.github.pylonmc.pylon.base.items.Hammer;
 import io.github.pylonmc.pylon.base.items.MonsterJerky;
 import io.github.pylonmc.pylon.base.items.*;
+import io.github.pylonmc.pylon.base.items.basichealing.*;
 import io.github.pylonmc.pylon.core.item.BasicItemSchema;
 import io.github.pylonmc.pylon.core.item.ItemStackBuilder;
 import io.github.pylonmc.pylon.core.item.PylonItemSchema;
@@ -15,8 +16,6 @@ import io.papermc.paper.datacomponent.item.Consumable;
 import io.papermc.paper.datacomponent.item.FoodProperties;
 import io.papermc.paper.datacomponent.item.ItemAttributeModifiers;
 import io.papermc.paper.datacomponent.item.PotionContents;
-import io.papermc.paper.datacomponent.item.UseCooldown;
-import io.papermc.paper.datacomponent.item.UseRemainder;
 import org.bukkit.Color;
 import org.bukkit.Material;
 import org.bukkit.NamespacedKey;
