+++ src/main/java/io/github/pylonmc/pylon/base/items/basichealing/Disinfectant.java
@@ -1,4 +1,4 @@
-package io.github.pylonmc.pylon.base.items;
+package io.github.pylonmc.pylon.base.items.basichealing;
 
 import io.github.pylonmc.pylon.core.item.PylonItemSchema;
 import io.github.pylonmc.pylon.core.item.SimplePylonItem;
@@ -9,18 +9,17 @@ import org.bukkit.inventory.ItemStack;
 import org.bukkit.inventory.ShapedRecipe;
 import org.bukkit.inventory.recipe.CraftingBookCategory;
 
-public class Plaster extends PylonItemSchema {
-    public Plaster(NamespacedKey id, Class<? extends SimplePylonItem> itemClass, ItemStack template){
+public class Disinfectant extends PylonItemSchema {
+    public Disinfectant(NamespacedKey id, Class<? extends SimplePylonItem> itemClass, ItemStack template){
         super(id, itemClass, template);
         ShapedRecipe craftingRecipe = new ShapedRecipe(id, template);
         craftingRecipe.shape(
-                "CC ",
-                "CC ",
-                "   "
+                "DDD",
+                "D D",
+                "DDD"
         );
-        craftingRecipe.setIngredient('C', Material.CLAY);
+        craftingRecipe.setIngredient('D', Material.DRIPSTONE_BLOCK);
         craftingRecipe.setCategory(CraftingBookCategory.MISC);
         RecipeTypes.VANILLA_CRAFTING.addRecipe(craftingRecipe);
     }
-
 }
