+++ src/main/java/io/github/pylonmc/pylon/base/items/basichealing/Plaster.java
@@ -1,6 +1,5 @@
-package io.github.pylonmc.pylon.base.items;
+package io.github.pylonmc.pylon.base.items.basichealing;
 
-import io.github.pylonmc.pylon.core.item.PylonItem;
 import io.github.pylonmc.pylon.core.item.PylonItemSchema;
 import io.github.pylonmc.pylon.core.item.SimplePylonItem;
 import io.github.pylonmc.pylon.core.recipe.RecipeTypes;
@@ -10,17 +9,18 @@ import org.bukkit.inventory.ItemStack;
 import org.bukkit.inventory.ShapedRecipe;
 import org.bukkit.inventory.recipe.CraftingBookCategory;
 
-public class Disinfectant extends PylonItemSchema {
-    public Disinfectant(NamespacedKey id, Class<? extends SimplePylonItem> itemClass, ItemStack template){
+public class Plaster extends PylonItemSchema {
+    public Plaster(NamespacedKey id, Class<? extends SimplePylonItem> itemClass, ItemStack template){
         super(id, itemClass, template);
         ShapedRecipe craftingRecipe = new ShapedRecipe(id, template);
         craftingRecipe.shape(
-                "DDD",
-                "D D",
-                "DDD"
+                "CC ",
+                "CC ",
+                "   "
         );
-        craftingRecipe.setIngredient('D', Material.DRIPSTONE_BLOCK);
+        craftingRecipe.setIngredient('C', Material.CLAY);
         craftingRecipe.setCategory(CraftingBookCategory.MISC);
         RecipeTypes.VANILLA_CRAFTING.addRecipe(craftingRecipe);
     }
+
 }
