+++ src/main/java/io/github/pylonmc/pylon/base/items/basichealing/Splint.java
@@ -1,4 +1,4 @@
-package io.github.pylonmc.pylon.base.items;
+package io.github.pylonmc.pylon.base.items.basichealing;
 
 import io.github.pylonmc.pylon.core.item.PylonItem;
 import io.github.pylonmc.pylon.core.item.PylonItemSchema;
@@ -16,33 +16,29 @@ import org.jetbrains.annotations.NotNull;
 import java.util.Objects;
 
 import static io.github.pylonmc.pylon.base.PylonItems.PLASTER;
-import static io.github.pylonmc.pylon.base.PylonItems.FIBER;
-import static io.github.pylonmc.pylon.base.PylonItems.DISINFECTANT;
 
-public class Medkit extends PylonItemSchema {
-    public static final double MEDKIT_HEAL_AMOUNT = 20;
-    public Medkit(NamespacedKey id, Class<? extends PylonItem<? extends PylonItemSchema>> itemClass, ItemStack template){
+public class Splint extends PylonItemSchema {
+    public static final double SPLINT_HEAL_AMOUNT = 12;
+    public Splint(NamespacedKey id, Class<? extends PylonItem<? extends PylonItemSchema>> itemClass, ItemStack template){
         super(id, itemClass, template);
         ShapedRecipe craftingRecipe = new ShapedRecipe(id, template);
         craftingRecipe.shape(
-                "PFP",
-                "DDD",
-                "PFP"
+                "PPP",
+                "   ",
+                "PPP"
         );
         craftingRecipe.setIngredient('P', PLASTER.getItemStack());
-        craftingRecipe.setIngredient('F', FIBER.getItemStack());
-        craftingRecipe.setIngredient('D', DISINFECTANT.getItemStack());
         craftingRecipe.setCategory(CraftingBookCategory.EQUIPMENT);
         RecipeTypes.VANILLA_CRAFTING.addRecipe(craftingRecipe);
     }
 
-    public static class Item extends PylonItem<Medkit> implements BlockInteractor {
-        public Item(Medkit schema, ItemStack itemStack) { super(schema, itemStack); }
+    public static class Item extends PylonItem<Splint> implements BlockInteractor {
+        public Item(Splint schema, ItemStack itemStack) { super(schema, itemStack); }
 
         @Override
         public void onUsedToRightClickBlock(@NotNull PlayerInteractEvent event) {
             event.setUseInteractedBlock(Event.Result.DENY);
-            event.getPlayer().heal(MEDKIT_HEAL_AMOUNT, EntityRegainHealthEvent.RegainReason.CUSTOM);
+            event.getPlayer().heal(SPLINT_HEAL_AMOUNT, EntityRegainHealthEvent.RegainReason.CUSTOM);
             event.getPlayer().getInventory().removeItem(Objects.requireNonNull(event.getItem()));
             event.getPlayer().updateInventory();
         }
